feat: implement engine exception hierarchy and move ViewportManager to manager package
