+++ src/main/java/net/luxsolari/engine/exceptions/ResourceInitializationException.java
@@ -4,7 +4,7 @@ package net.luxsolari.engine.exceptions;
  * Exception thrown when resources cannot be properly initialized.
  * This typically occurs during system startup or resource loading.
  */
-public class ResourceInitializationException extends RuntimeException {
+public class ResourceInitializationException extends EngineException {
   
   /**
    * Creates a new ResourceInitializationException with a message and cause.
