+++ src/main/resources/logging.properties
@@ -9,3 +9,17 @@ java.util.logging.ConsoleHandler.formatter=java.util.logging.SimpleFormatter
 
 # Default level for the console handler
 java.util.logging.ConsoleHandler.level=ALL
+
+# Package-level logging configuration
+# Engine subsystems - INFO level for operational visibility
+net.luxsolari.engine.systems.level=INFO
+net.luxsolari.engine.manager.level=INFO
+
+# ECS system - WARNING level to reduce noise
+net.luxsolari.engine.ecs.level=WARNING
+
+# Game states - INFO level for game flow tracking
+net.luxsolari.game.states.level=INFO
+
+# Exception handling - SEVERE level only (exceptions are logged at throw time)
+net.luxsolari.engine.exceptions.level=SEVERE
